Refactor book API views and update URL routing for clarity

+++ book_api/views.py
@@ -6,42 +6,40 @@ from book_api.serializer import BookSerializer
 from rest_framework import status
 
 # Create your views here.
-@api_view(['GET'])
-def book_list(request):
-    books = Book.objects.all()
-    serializer = BookSerializer(books, many=True)
-    return Response(serializer.data)
-
-@api_view(["POST"])
-def book_create(request):
-    serializer = BookSerializer(data=request.data)
-    if serializer.is_valid():
-        serializer.save()
+@api_view(['GET','POST'])
+def books(request):
+    if request.method == "GET":
+        books = Book.objects.all()
+        serializer = BookSerializer(books, many=True)
         return Response(serializer.data)
-    else:
-        return Response(serializer.errors)
 
-@api_view(['GET'])    
+    if request.method == "POST":
+        serializer = BookSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data,status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors)
+
+@api_view(['GET','PUT','DELETE'])    
 def book(request, id):
-    try:
-        book = Book.objects.get(pk=id)
-        serializer = BookSerializer(book)
-        return Response(serializer.data)
-    except:
-        return Response({"error":"Eşleşen bir kayıt bulunamadı."}, status=status.HTTP_404_NOT_FOUND)
-    
-@api_view(["PUT"])
-def book_update(request, id):
     book = Book.objects.get(pk=id)
-    seralizer = BookSerializer(book,data = request.data)
-    if seralizer.is_valid():
-        seralizer.save()
-        return Response(seralizer.data)
-    else:
-        return Response(seralizer.errors)
+    if request.method == "GET":
+        try:
+            
+            serializer = BookSerializer(book)
+            return Response(serializer.data)
+        except:
+            return Response({"error":"Eşleşen bir kayıt bulunamadı."}, status=status.HTTP_404_NOT_FOUND)
     
-@api_view(["DELETE"])
-def book_delete(request, id):
-    book = Book.objects.get(pk=id)
-    book.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    if request.method == "PUT":
+        seralizer = BookSerializer(book,data = request.data)
+        if seralizer.is_valid():
+            seralizer.save()
+            return Response(seralizer.data)
+        else:
+            return Response(seralizer.errors)
+        
+    if request.method == "DELETE":
+        book.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
